+++ core/src/main/java/react4j/annotations/View.java
@@ -13,6 +13,7 @@ import javax.annotation.Nonnull;
 @Documented
 @Target( ElementType.TYPE )
 @ActAsStingComponent
+@ActAsArezComponent
 public @interface View
 {
   /**