+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Warn when `immutable = true` is specified on method-based `@Input` declarations and allow the warning to be suppressed while downstream projects migrate.
+* Annotate `@View` with `@ActAsArezComponent` to ensure it is valid to use arez annotations on react components.
 
 ### [v0.210](https://github.com/react4j/react4j/tree/v0.210) (2026-04-15) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.209...v0.210)