+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.211](https://github.com/react4j/react4j/tree/v0.211) (2026-04-15) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.210...v0.211)
+
+Changes in this release:
 
 * Warn when `immutable = true` is specified on method-based `@Input` declarations and allow the warning to be suppressed while downstream projects migrate.
 * Annotate `@View` with `@ActAsArezComponent` to ensure it is valid to use arez annotations on react components.